[황민영] ⚡ perf : 컴포넌트 분리
[황민영] ⚡ perf : 컴포넌트 분리
